ios/options: helper for key-value options

+++ ios/IJKMediaPlayer/IJKMediaPlayer/IJKFFMoviePlayerController+OptionConvenience.h
@@ -29,10 +29,10 @@
 - (void)setSwsOptionValue:          (NSString *)value forKey:(NSString *)key;
 - (void)setPlayerOptionValue:       (NSString *)value forKey:(NSString *)key;
 
-- (void)setFormatOptionIntValue:    (NSInteger)value forKey:(NSString *)key;
-- (void)setCodecOptionIntValue:     (NSInteger)value forKey:(NSString *)key;
-- (void)setSwsOptionIntValue:       (NSInteger)value forKey:(NSString *)key;
-- (void)setPlayerOptionIntValue:    (NSInteger)value forKey:(NSString *)key;
+- (void)setFormatOptionIntValue:    (int64_t)value forKey:(NSString *)key;
+- (void)setCodecOptionIntValue:     (int64_t)value forKey:(NSString *)key;
+- (void)setSwsOptionIntValue:       (int64_t)value forKey:(NSString *)key;
+- (void)setPlayerOptionIntValue:    (int64_t)value forKey:(NSString *)key;
 
 - (void)setMaxBufferSize:(int)maxBufferSize; // -1 for default size
